# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SCO Lead Management API
  version: 1.0.0
  description: >
    API untuk mengelola data EO/Mitra (Lead), assignment PIC, tracking status pipeline,
    dan dokumentasi aktivitas follow-up.
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ContactPerson:
      type: object
      properties:
        id:
          type: string
          format: uuid
        leadId:
          type: string
          format: uuid
        nama:
          type: string
        noTelp:
          type: string
        jabatan:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Assignment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        leadId:
          type: string
          format: uuid
        picId:
          type: string
          format: uuid
        isActive:
          type: boolean
        assignedAt:
          type: string
          format: date-time
        reassignedAt:
          type: string
          format: date-time
          nullable: true
        reassignedNotes:
          type: string
          nullable: true
        assignedBy:
          type: string
          format: uuid
...
    CreateContactRequest:
      type: object
      required: [nama, no_telp, jabatan]
      properties:
        nama:
          type: string
          minLength: 1
          maxLength: 255
        no_telp:
          type: string
          minLength: 1
          maxLength: 20
        jabatan:
          type: string
          minLength: 1
          maxLength: 100
    UpdateContactRequest:
      type: object
      properties:
        nama:
          type: string
          maxLength: 255
        no_telp:
          type: string
          maxLength: 20
        jabatan:
          type: string
          maxLength: 100
    AssignRequest:
      type: object
      required: [pic_id]
      properties:
        pic_id:
          type: string
          format: uuid
        notes:
          type: string
          description: If provided, treated as a re-assignment
...
    LoginRequest:
      type: object
      required: [employee_id]
      properties:
        employee_id:
          type: string
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    ApiError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
  responses:
    Unauthorized:
      description: Token tidak valid atau tidak ditemukan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Token tidak valid atau tidak ditemukan
    Forbidden:
      description: Akses ditolak
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Anda tidak memiliki akses untuk operasi ini
    NotFound:
      description: Resource tidak ditemukan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource tidak ditemukan
    ValidationError:
      description: Data tidak valid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Data tidak valid
              details: {}
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login dengan employee_id
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                      refresh_token:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== Users ====================
...
  /api/leads/{leadId}/contacts:
    parameters:
      - name: leadId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Contacts]
      summary: Tambah Contact Person
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContactRequest'
      responses:
        '201':
          description: Contact berhasil ditambahkan
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ContactPerson'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [Contacts]
      summary: Daftar Contact Person per Lead
      responses:
        '200':
          description: Daftar contact
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContactPerson'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/leads/{leadId}/contacts/{id}:
    parameters:
      - name: leadId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Contacts]
      summary: Update Contact Person
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContactRequest'
      responses:
        '200':
          description: Contact berhasil diupdate
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ContactPerson'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== Assignments ====================
  /api/leads/{leadId}/assignments:
    parameters:
      - name: leadId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Assignments]
      summary: Assign atau Re-assign Lead ke PIC (Superior only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRequest'
      responses:
        '201':
          description: Assignment berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Assignment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags: [Assignments]
      summary: Histori assignment (Superadmin, Superior)
      responses:
        '200':
          description: Histori assignment
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Assignment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==================== Status Pipeline ====================
...
  /api/leads/{leadId}/activities/{activityId}/evidence:
    parameters:
      - name: leadId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: activityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Evidence]
      summary: Upload evidence file (PIC only)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [evidence]
              properties:
                evidence:
                  type: string
                  format: binary
                  description: File JPG/PNG, max 5MB
      responses:
        '200':
          description: Evidence berhasil diupload
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      evidencePath:
                        type: string
                      filename:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          description: File terlalu besar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                success: false
                error:
                  code: FILE_TOO_LARGE
                  message: Ukuran file maksimal adalah 5MB
        '415':
          description: Format file tidak didukung
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                success: false
                error:
                  code: INVALID_FILE_TYPE
                  message: File harus berformat JPG atau PNG
  /api/evidence/{filename}:
    parameters:
      - name: filename
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Evidence]
      summary: Download/view evidence file
      responses:
        '200':
          description: File evidence
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== Audit Trail ====================
...